Refactor hue
