+++ plugin/src/main/java/com/craftmend/openaudiomc/generic/networking/packets/PacketClientApplyHueColor.java
@@ -1,6 +1,6 @@
 package com.craftmend.openaudiomc.generic.networking.packets;
 
-import com.craftmend.openaudiomc.generic.objects.SerializedHueColor;
+import com.craftmend.openaudiomc.generic.hue.SerializedHueColor;
 import com.craftmend.openaudiomc.generic.networking.abstracts.AbstractPacket;
 import com.craftmend.openaudiomc.generic.networking.enums.PacketChannel;
 import com.craftmend.openaudiomc.generic.networking.payloads.HueColorPayload;
