+++ plugin/src/main/java/com/craftmend/openaudiomc/generic/networking/payloads/HueColorPayload.java
@@ -1,6 +1,6 @@
 package com.craftmend.openaudiomc.generic.networking.payloads;
 
-import com.craftmend.openaudiomc.generic.objects.SerializedHueColor;
+import com.craftmend.openaudiomc.generic.hue.SerializedHueColor;
 import com.craftmend.openaudiomc.generic.networking.abstracts.AbstractPacketPayload;
 import lombok.AllArgsConstructor;
 import lombok.Data;
